Splitting reset-style-per-node
