+++ lib/assets/javascripts/cartodb3/data/user-actions.js
@@ -7,6 +7,7 @@ var camshaftReference = require('./camshaft-reference');
 var layerTypesAndKinds = require('./layer-types-and-kinds');
 var layerColors = require('./layer-colors');
 var Notifier = require('../components/notifier/notifier');
+var ResetStylePerNode = require('../helpers/reset-style-per-node');
 
 /**
  * Coordinate side-effects done on explicit interactions.
@@ -119,87 +120,6 @@ module.exports = function (params) {
     }
   };
 
-  var resetStylesWhenGeometryChanges = function (nodeDefModel, layerDefModel, forceStyleUpdate) {
-    var CHANGE_READY = 'change:ready';
-    var onQueryGeometryAndSchemaReady;
-    var styleModel = layerDefModel.styleModel;
-
-    onQueryGeometryAndSchemaReady = function () {
-      var queryGeometryModel = nodeDefModel.queryGeometryModel;
-      var querySchemaModel = nodeDefModel.querySchemaModel;
-
-      if (queryGeometryModel.get('ready') && querySchemaModel.get('ready')) {
-        queryGeometryModel.unbind(CHANGE_READY, onQueryGeometryAndSchemaReady);
-        querySchemaModel.unbind(CHANGE_READY, onQueryGeometryAndSchemaReady);
-      } else {
-        return; // wait until ready
-      }
-
-      var CHANGE_STATUS = 'change:status';
-      var saveDefaultStylesIfStillRelevant;
-
-      saveDefaultStylesIfStillRelevant = function () {
-        var unbindNodeChangeStatus = function () {
-          queryGeometryModel.unbind(CHANGE_STATUS, saveDefaultStylesIfStillRelevant);
-          querySchemaModel.unbind(CHANGE_STATUS, saveDefaultStylesIfStillRelevant);
-        };
-
-        var applyStylesAndSave = function () {
-          var simpleGeometryType = queryGeometryModel.get('simple_geom');
-          if (simpleGeometryType) {
-            styleModel.setDefaultPropertiesByType('simple', simpleGeometryType);
-          } else {
-            styleModel.setDefaultPropertiesByType('none'); // fallback if there is no known geometry
-          }
-          layerDefModel.save();
-        };
-
-        if (queryGeometryModel.isFetching() || querySchemaModel.isFetching()) { // Wait until fetched
-          return;
-        } else if (!queryGeometryModel.isFetching() && !queryGeometryModel.isFetched() &&
-          !querySchemaModel.isFetching() && !querySchemaModel.isFetched()) { // Node status has been resetted
-          unbindNodeChangeStatus();
-          return;
-        } else { // Fetched!
-          unbindNodeChangeStatus();
-        }
-
-        // If the geometry doesn't change and all columns used for style-def-model are still present + the styles are not custom
-        // we can reset the style form with the geometry provided
-        var haveDifferentSchema = querySchemaModel.hasDifferentSchemaThan(styleModel.getColumnsUsedForStyle());
-        if (
-            !forceStyleUpdate && (
-              (!queryGeometryModel.hasChanged('simple_geom') && !haveDifferentSchema) ||
-              layerDefModel.get('custom_cartocss')
-            )
-          ) {
-          return;
-        }
-
-        if ( // Only apply changes if:
-          (
-            layerDefinitionsCollection.contains(layerDefModel) && // layer still exist
-            layerDefModel.get('source') === nodeDefModel.id // node is still the head of layer
-          ) || forceStyleUpdate // we force the style update
-        ) {
-          applyStylesAndSave();
-        }
-      };
-
-      saveDefaultStylesIfStillRelevant();
-
-      queryGeometryModel.bind(CHANGE_STATUS, saveDefaultStylesIfStillRelevant);
-      querySchemaModel.bind(CHANGE_STATUS, saveDefaultStylesIfStillRelevant);
-
-      queryGeometryModel.fetch();
-      querySchemaModel.fetch();
-    };
-
-    nodeDefModel.queryGeometryModel.bind(CHANGE_READY, onQueryGeometryAndSchemaReady);
-    nodeDefModel.querySchemaModel.bind(CHANGE_READY, onQueryGeometryAndSchemaReady);
-    onQueryGeometryAndSchemaReady();
-  };
-
   return {
     saveAnalysis: function (analysisFormModel) {
       if (!analysisFormModel.isValid()) return;
@@ -248,7 +168,7 @@ module.exports = function (params) {
         }
       });
 
-      resetStylesWhenGeometryChanges(nodeDefModel, layerDefModel);
+      ResetStylePerNode(nodeDefModel, layerDefModel);
 
       return nodeDefModel;
     },
@@ -260,7 +180,7 @@ module.exports = function (params) {
       nodeDefModel.set('query', query || '');
 
       analysisDefinitionsCollection.saveAnalysisForLayer(layerDefModel);
-      resetStylesWhenGeometryChanges(nodeDefModel, layerDefModel);
+      ResetStylePerNode(nodeDefModel, layerDefModel);
       layerDefModel.save(); // to make sure layer's source ref is persisted
     },
 
@@ -325,7 +245,7 @@ module.exports = function (params) {
         }
       });
 
-      resetStylesWhenGeometryChanges(primarySourceNode, layerDefModel);
+      ResetStylePerNode(primarySourceNode, layerDefModel);
 
       deleteOrphanedAnalyses();
     },
@@ -420,7 +340,7 @@ module.exports = function (params) {
       var affectedWidgetAttrsBySourceChange = [];
 
       var onNewLayerSaved = function (layer) {
-        resetStylesWhenGeometryChanges(layer.getAnalysisDefinitionNodeModel(), layer, true);
+        ResetStylePerNode(layer.getAnalysisDefinitionNodeModel(), layer, true);
         layerDefinitionsCollection.save(); // to persist layers order
       };
 
@@ -520,7 +440,7 @@ module.exports = function (params) {
           layerDefinitionsCollection.add(prevLayer, {at: newPosition});
 
           // Reset styles from previous layer
-          resetStylesWhenGeometryChanges(prevLayer.getAnalysisDefinitionNodeModel(), prevLayer, true);
+          ResetStylePerNode(prevLayer.getAnalysisDefinitionNodeModel(), prevLayer, true);
 
           newLayerDefModel.save(null, {
             success: function () {
